Millisecond timeouts are now long to avoid overflow.

+++ trunk/java/server/src/main/org/avis/federation/Acceptor.java
@@ -88,9 +88,8 @@ public class Acceptor implements IoHandler, Closeable
     this.options = options;
     this.links = new HashSet<Link> ();
     
-    // todo this should be a long to avoid overflow
-    int requestTimeout = options.getInt ("Federation.Request-Timeout") * 1000;
-    int keepaliveInterval = 
+    long requestTimeout = options.getInt ("Federation.Request-Timeout") * 1000;
+    long keepaliveInterval = 
       options.getInt ("Federation.Keepalive-Interval") * 1000;
     
     SocketAcceptorConfig acceptorConfig = new SocketAcceptorConfig ();
